Remove dimensions reducer file
