+++ src/app/services/actions/dimensions-action-creator.spec.ts
@@ -26,7 +26,8 @@ describe('Dimensions setting on GE', () => {
 
     const expectedActions = {
       type: RESIZE_SUCCESS,
-      response: dimensions
+      payload:dimensions,
+      response: dimensions // TODO: remove
     }
 
     // Act
