+++ src/app/services/actions/dimensions-action-creator.ts
@@ -5,6 +5,7 @@ import { RESIZE_SUCCESS } from '../redux-constants';
 export function setDimensions(response: IDimensions): AppAction {
   return {
     type: RESIZE_SUCCESS,
-    response
+    payload: response,
+    response // TODO: remove after RTK migration is done
   };
 }
