+++ src/app/services/reducers/index.ts
@@ -5,10 +5,11 @@ import sampleQuery from '../slices/sample-query.slice';
 import graphExplorerMode from '../slices/explorer-mode.slice';
 import authReducer from '../slices/auth.slice';
 import autoComplete from '../slices/autocomplete.slice';
+import dimensions from '../slices/dimensions.slice';
 
 import { collections } from './collections-reducer';
 import { devxApi } from './devxApi-reducers';
-import { dimensions } from './dimensions-reducers';
+// import { dimensions } from './dimensions-reducers';
 import { scopes } from './permissions-reducer';
 import { profile } from './profile-reducer';
 import { proxyUrl } from './proxy-url-reducer';
