+++ src/app/services/redux-constants.ts
@@ -36,7 +36,7 @@ export const REMOVE_ALL_HISTORY_ITEMS_SUCCESS = 'REMOVE_ALL_HISTORY_ITEMS_SUCCES
 export const AUTOCOMPLETE_FETCH_SUCCESS = 'autoComplete/fetch/fulfilled';
 export const AUTOCOMPLETE_FETCH_ERROR = 'autoComplete/fetch/rejected';
 export const AUTOCOMPLETE_FETCH_PENDING = 'autoComplete/fetch/pending';
-export const RESIZE_SUCCESS = 'RESIZE_SUCCESS';
+export const RESIZE_SUCCESS = 'dimensions/resizeSuccess';
 export const RESPONSE_EXPANDED = 'RESPONSE_EXPANDED';
 export const PERMISSIONS_PANEL_OPEN = 'PERMISSIONS_PANEL_OPEN';
 export const AUTHENTICATION_PENDING = 'auth/setAuthenticationPending';
